Crate rs_es

A client for ElasticSearch's REST API

The Client itself is used as the central access point, from which numerous operations are defined implementing each of the specific ElasticSearch APIs.

Warning: at the time of writing the majority of such APIs are currently unimplemented.

Modules

error
Errors and error conversion code for the rs_es crate

operations
Implementations of specific ElasticSearch operations

query
Implementation of the ElasticSearch Query DSL.

units
Various re-occuring types that are used by the ElasticSearch API.

util
Miscellaneous code used in numerous places

Structs

Client
The core of the ElasticSearch client, owns a HTTP connection.

Functions

do_req
Process the result of an HTTP request, returning the status code and the Json result (if the result had a body) or an EsError if there were any errors