Function rs_es::do_req

pub fn do_req(resp: &mut Response) -> Result<(StatusCode, Option<Json>), EsError>

Process the result of an HTTP request, returning the status code and the Json result (if the result had a body) or an EsError if there were any errors

This function is exposed to allow extensions to certain operations, it is not expected to be used by consumers of the library