Module rs_es::operations

Implementations of specific ElasticSearch operations

The various methods on Client are entry points to ElasticSearch's set of operations. This module, and it's child modules are the implementation of those operations.

Modules

bulk
Implementation of the Bulk API

delete
Implementation of delete operations, both Delete-By-Query and Delete-By-Id

get
Implementation of the Get API

index
Implementation of ElasticSearch Index operation

search
Implementations of both Search-by-URI and Search-by-Query operations

Structs

RefreshOperation

RefreshResult
Result of a refresh request

ShardCountResult
Shared struct for operations that include counts of success/failed shards. This is returned within various other result structs.