Struct rs_es::operations::mapping::MappingOperation

pub struct MappingOperation<'a, 'b> {
    // some fields omitted
}

An indexing operation

Methods

impl<'a, 'b> MappingOperation<'a, 'b>

fn new(client: &'a mut Client, index: &'b str) -> MappingOperation<'a, 'b>

fn with_mapping(&'b mut self, mapping: &'b Mapping) -> &'b mut Self

Set the actual mapping

fn with_settings(&'b mut self, settings: &'b Settings) -> &'b mut Self

Set the settings

fn send(&'b mut self) -> Result<MappingResult, EsError>

If settings have been provided, the index will be created with them. If the index already exists, an Err(EsError) will be returned. If mapping have been set too, the properties will be applied. The index will be unavailable during this process. Nothing will be done if either mapping and settings are not present.