Module rs_es::operations::search

Implementations of both Search-by-URI and Search-by-Query operations

Modules

aggregations
Implementation of ElasticSearch aggregations

Structs

GeoDistance
Representing sort options for sort by geodistance

ScanIterator

ScanResult
Used when scanning and scrolling through results, a ScanResult is returned when the scan is opened. To scroll through the results an application has two options:

Script
Representing options for sort by script

SearchHitsHitsResult

SearchHitsResult

SearchQueryOperation

SearchResult

SearchURIOperation
Representing a search-by-uri option

Sort
A full sort clause

SortField
Representing sort options for a specific field, can be combined with others to produce the full sort clause

Enums

Missing
Options for handling (missing values)[https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-sort.html#_missing_values]

Mode
The (Sort mode option)[https://www.elastic.co/guide/en/elasticsearch/reference/current/search-request-sort.html#_sort_mode_option].

Order
Order of a sort

SearchType
Options for the various search_type parameters

SortBy

Source
Options for source filtering