Module rs_es::operations::search::aggregations

Implementation of ElasticSearch aggregations

Structs

Aggregations
The set of aggregations

AggregationsResult

Avg
Avg aggregation

AvgResult

Bounds
Used by the ExtendedStatsResult

Cardinality
Cardinality aggregation

CardinalityResult

Children
Children aggregation - sub-aggregations run against the child document

ChildrenResult

DateHistogram

DateHistogramBucketResult

DateHistogramResult

DateRange
Date range aggregation. See: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-daterange-aggregation.html

DateRangeBucketResult

DateRangeInst
A specific element of a range for a DateRange aggregation

DateRangeResult

ExtendedBounds
Histogram aggregation.

ExtendedStats
Extended stats aggregation

ExtendedStatsResult

Filter
Filter aggregation

FilterResult

Filters
Filters aggregation

FiltersBucketResult

FiltersResult

GeoBounds
Geo Bounds aggregation

GeoBoundsResult

GeoDistance

GeoDistanceBucketResult

GeoDistanceInst

GeoDistanceResult

GeoHash
Geohash aggregation

GeoHashBucketResult

GeoHashResult

Global
Global aggregation, defines a single global bucket. Can only be used as a top-level aggregation. See: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-global-aggregation.html

GlobalResult

Histogram

HistogramBucketResult
Used for histogram results

HistogramResult

Max
Max aggregation

MaxResult

Min
Min aggregation

MinResult

Missing
Missing aggregation

MissingResult

Nested
Nested aggregation

NestedResult

Order
Used to define the ordering of buckets in a some bucketted aggregations

PercentileRanks
Percentile Ranks aggregation

PercentileRanksResult

Percentiles
Percentiles aggregation, see: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-metrics-percentile-aggregation.html

PercentilesResult

Range
Range aggregations

RangeBucketResult

RangeInst
A specific range, there will be many of these making up a range aggregation

RangeResult

ReverseNested
Reverse nested aggregation, will produce an error if used anywhere other than inside a nested aggregation.

ReverseNestedResult

Script
Script attributes for various attributes

ScriptedMetric
Scripted method aggregation

ScriptedMetricResult

Stats
Stats aggregation

StatsResult

Sum
Sum aggregation

SumResult

Terms
Terms aggregation

TermsBucketResult

TermsResult

ValueCount
Value count aggregation

ValueCountResult

Enums

Aggregation
Aggregations are either metrics or bucket-based aggregations

AggregationResult
The result of one specific aggregation

BucketAggregation
The set of bucket aggregations

FieldOrScript
A common pattern is for an aggregation to accept a field or a script

Interval

MetricsAggregation
Individual aggregations and their options

OrderKey
Order - used for some bucketing aggregations to determine the order of buckets

Scripts

TimeZone
Date histogram and related fields