Module rs_es::operations::search::aggregations::bucket

Bucket-based aggregations

Structs

Children
Children aggregation - sub-aggregations run against the child document

ChildrenResult

DateHistogram

DateHistogramBucketResult

DateHistogramResult

DateRange
Date range aggregation. See: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-daterange-aggregation.html

DateRangeBucketResult

DateRangeInner

DateRangeInst
A specific element of a range for a DateRange aggregation

DateRangeResult

ExtendedBounds
Histogram aggregation.

Filter
Filter aggregation

FilterResult
Filter result

Filters
Filters aggregation

FiltersBucketResult

FiltersResult

GeoDistance

GeoDistanceBucketResult

GeoDistanceInst

GeoDistanceResult

GeohashGrid
Geohash aggregation

GeohashGridBucketResult

GeohashGridResult

Global
Global aggregation, defines a single global bucket. Can only be used as a top-level aggregation. See: https://www.elastic.co/guide/en/elasticsearch/reference/current/search-aggregations-bucket-global-aggregation.html

GlobalResult
Global result

Histogram

HistogramBucketResult
Used for histogram results

HistogramResult

Missing
Missing aggregation

MissingResult

Nested
Nested aggregation

NestedResult

Order
Used to define the ordering of buckets in a some bucketted aggregations

Range
Range aggregations

RangeBucketResult

RangeInner

RangeInst
A specific range, there will be many of these making up a range aggregation

RangeResult

ReverseNested
Reverse nested aggregation, will produce an error if used anywhere other than inside a nested aggregation.

ReverseNestedResult

Terms
Terms aggregation

TermsBucketResult

TermsInner

TermsResult
Terms result

Enums

BucketAggregation
The set of bucket aggregations

BucketAggregationResult

ExecutionHint

Interval

OrderKey
Order - used for some bucketing aggregations to determine the order of buckets

TimeZone
Date histogram and related fields