Struct rs_es::operations::search::highlight::Highlight

pub struct Highlight {
    pub fields: HashMap<String, Setting>,
    pub pre_tags: Option<Vec<String>>,
    pub post_tags: Option<Vec<String>>,
    pub encoder: Option<Encoders>,
}

Fields
fields: HashMap<String, Setting> pre_tags: Option<Vec<String>> post_tags: Option<Vec<String>> encoder: Option<Encoders>
Methods

impl Highlight

fn new() -> Highlight

Create an Highlight entity without any field or setting specified as they are supposed to be added via the add method.

Examples

use rs_es::operations::search::highlight::{Highlight, Setting, SettingTypes, Encoders};

let mut highlight = Highlight::new().with_encoder(Encoders::HTML).to_owned();
let setting = Setting::new().with_type(SettingTypes::Plain).to_owned();
highlight.add_setting("first_name".to_owned(), setting);

fn with_encoder(&mut self, encoder: Encoders) -> &mut Highlight

fn with_pre_tags(&mut self, pre_tags: Vec<String>) -> &mut Highlight

fn with_post_tags(&mut self, post_tags: Vec<String>) -> &mut Highlight

fn add_setting(&mut self, name: String, setting: Setting)

Add a field to highlight to the set