Struct rs_es::operations::search::ScanResultInterim

pub struct ScanResultInterim<T: Deserialize> {
    // some fields omitted
}

Used when scanning and scrolling through results, a ScanResult is returned when the scan is opened. To scroll through the results an application has two options:

Call scroll repeatedly until the returned results have zero hits. If this approach is taken, the caller is also responsible for calling close when finished, to prevent any server-side resources being held open.

Call 'iter' to create an iterator from which the hits can be read. If this approach is taken, there is no need to call close, it will be called automatically when iteration ends.

See also the official ElasticSearch documentation for proper use of this functionality.