Struct rs_es::operations::search::SearchQueryOperation

pub struct SearchQueryOperation<'a, 'b> {
    // some fields omitted
}

Methods

impl<'a, 'b> SearchQueryOperation<'a, 'b>

fn new(client: &'a mut Client) -> SearchQueryOperation<'a, 'b>

fn with_indexes(&'b mut self, indexes: &'b [&'b str]) -> &'b mut Self

fn with_types(&'b mut self, doc_types: &'b [&'b str]) -> &'b mut Self

fn with_query(&'b mut self, query: &'b Query) -> &'b mut Self

fn with_timeout(&'b mut self, timeout: &'b str) -> &'b mut Self

fn with_from(&'b mut self, from: u64) -> &'b mut Self

fn with_size(&'b mut self, size: u64) -> &'b mut Self

fn with_terminate_after(&'b mut self, terminate_after: u64) -> &'b mut Self

fn with_stats<S>(&'b mut self, stats: &[S]) -> &'b mut Self where S: ToString

fn with_min_score(&'b mut self, min_score: f64) -> &'b mut Self

fn with_sort(&'b mut self, sort: &'b Sort) -> &'b mut Self

fn with_track_scores(&'b mut self, track_scores: bool) -> &'b mut Self

fn with_source(&'b mut self, source: Source<'b>) -> &'b mut Self

Specify source filtering, by default full source will be returned in a hit

To switch-off source document in each hit: with_source(Source::Off). To include fields: with_source(Source::include(&["field_name"])), To exclude fields: with_source(Source::exclude(&["field_name"])), To include and exclude: with_source(Source::filter(&["include"], &["exclude"]))

fn with_aggs(&'b mut self, aggs: &'b Aggregations) -> &'b mut Self

Specify any aggregations

fn with_routing<T: Into<OptionVal>>(&'a mut self, val: T) -> &'a mut Self

fn with_search_type<T: Into<OptionVal>>(&'a mut self, val: T) -> &'a mut Self

fn with_query_cache<T: Into<OptionVal>>(&'a mut self, val: T) -> &'a mut Self

fn send(&'b mut self) -> Result<SearchResult, EsError>

Performs the search with the specified query and options

fn scan(&'b mut self, scroll: Duration) -> Result<ScanResult, EsError>

Begins a scan with the specified query and options