Struct rs_es::operations::search::SortField

pub struct SortField {
    // some fields omitted
}

Representing sort options for a specific field, can be combined with others to produce the full sort clause

Methods

impl SortField

fn new<S: Into<String>>(field: S, order: Option<Order>) -> SortField

Create a SortField for a given field and order

fn with_mode<T: Into<Mode>>(self, val: T) -> Self

fn with_nested_path<T: Into<String>>(self, val: T) -> Self

fn with_nested_filter<T: Into<Filter>>(self, val: T) -> Self

fn with_missing<T: Into<Missing>>(self, val: T) -> Self

fn with_unmapped_type<T: Into<String>>(self, val: T) -> Self

fn build(self) -> SortBy