Module rs_es::units

Various re-occuring types that are used by the ElasticSearch API.

E.g. Duration

This isn't all types. Types that are specific to one API are defined in the appropriate place, e.g. types only used by the Query DSL are in query.rs

Structs

Duration
A time-period unit, will be formatted into the ElasticSearch standard format

Enums

DistanceType
DistanceType

DistanceUnit
DistanceUnit

DurationUnit
The units by which duration is measured.

GeoBox
Representing a geographic box

JsonVal
A Json value that's not a structural thing - i.e. just String, i64 and f64, no array or object

Location
Representing a geographic location

OneOrMany
A non-specific holder for an option which can either be a single thing, or multiple instances of that thing.