Trait rs_es::util::StrJoin

pub trait StrJoin {
    fn join(self, join: &str) -> String;
}

A custom String-join trait as the stdlib one is currently marked as unstable.

Required Methods

fn join(self, join: &str) -> String

Join an iterator of things that can be referenced as strings into a single owned-string by the given joining string

Example

use rs_es::util::StrJoin;

let data = vec!["a", "b", "c", "d"];
assert_eq!("a-b-c-d", data.iter().join("-"));

This will print: a-b-c-d